Example of Tutorial Gameplay (Beginner: Scales[CM])
THE CONCEPT:  The screen has two piano keyboards with note names visible
on each key.  The keyboard on the right is the computer, which  plays
notes in sequence expecting the player to press or click the
corresponding keys on their own keyboard. The computer also plays
automated "intro" music and "celebration" music when the player
successfully imitates the notes.  If the player hits the wrong note, on
the following beat the computer will play "reset" music and start the
drill over(see: Screwup.wav).  This example has three total drills, the
final one being the whole scale up and down.  (The higher keyboard is
the computer, the lower one is the player)

Added note:  It would  be great if a staff could also be visible
above/below the keyboards so that when each key is pressed, the
corresponding note appears on the staff for the duration that the note
is played.