Example of Tutorial Gameplay (Beginner: Rhythm)
There are two timpani drums on the screen and the higher-pitched one is
the computer, whereas the lower-pitched one is the player.  This is the
same concept as the Scales except it only focuses on reading rhythms.
The player won't have to worry about pitch, but unlike the scale drills,
the percussion drills will show rests so the players will be exposed to
the concepts of the note/rest dichotomy in this drill.

Of course all of these ideas are purely conceptual and if they don't
fit, there are other solutions.